Sequantial TASEP Coalescence Times sigma^(o|x) & sigma^(x|o)
ScritpN launches a new function in parall_tasep that initialize the
sigma_s and computes the Coalescence Time according to a SEQUENTIAL
update rule.
Few output npys files are stored in the folders according to the
probability.
LinRegLogLog reads these outputs and plots them on a LogLog scale
showing that the MEAN value of CT grows linearly and the MAX with a
factor approx 5/2.